% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_autoppt.R
\name{autoppt_format}
\alias{autoppt_format}
\title{Rmarkdown Document Format For Automated Powerpoint Presentations}
\usage{
autoppt_format(base_ppt = NULL, ...)
}
\arguments{
\item{base_ppt}{Path to the template presentation}

\item{...}{Other arguments to html_document}
}
\value{
An output format to be used in .Rmd documents
}
\description{
Rmarkdown Document Format For Automated Powerpoint Presentations
}
