% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_data.R
\name{format_chart_data}
\alias{format_chart_data}
\title{Process data for plotting}
\usage{
format_chart_data(
  data,
  categories = NULL,
  series = NULL,
  date_fmt = "mm/yy",
  return_all = FALSE,
  clean_categories = FALSE,
  space_categories_at = FALSE
)
}
\arguments{
\item{data}{data frame}

\item{categories}{column names (or indexes) which will be mapped as categories (e.g., x-axis)}

\item{series}{column names (or indexes) which will be mapped as series (e.g., y-axis)}

\item{date_fmt}{Excel data format string (e.g., "mm/yy")}

\item{return_all}{Should columns besides categories and series be returned}

\item{clean_categories}{Whether repeating categories should be replaced with NA (Excel usually needs this)}

\item{space_categories_at}{After which category level to add blank rows}
}
\value{
A formatted \code{chart_data} object with set \code{category}, \code{series}, \code{data_fmt} attributes
}
\description{
Process data for plotting
}
