% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.autopres-package.R
\docType{package}
\name{r.autopres-package}
\alias{r.autopres}
\alias{r.autopres-package}
\title{r.autopres: Tools for automating PowerPoint presentations}
\description{
.
}
\author{
\strong{Maintainer}: Arturs Janis Nikitins \email{ajnikitins@gmail.com}

}
\keyword{internal}
