% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_data.R
\name{replace_chart_data}
\alias{replace_chart_data}
\title{Replace PowerPoint chart data}
\usage{
replace_chart_data(input_ppt, output_ppt, slide_index, chart_index, chart_data)
}
\arguments{
\item{input_ppt}{Path to input presentation}

\item{output_ppt}{Path to output presentation}

\item{slide_index}{Index of slide}

\item{chart_index}{Index of chart}

\item{chart_data}{Data frame with new chart data (preferably \code{chart_data})}
}
\description{
Replace PowerPoint chart data
}
